web/refactor: get fallback
