+++ web/app/types/tendenzApiTypes.tsx
@@ -1,0 +1,18 @@
+export interface tendenzApiSigmaYesterdayDay {
+  close: number;
+  logReturn: number;
+  date: string;
+}
+
+export interface tendenzApiSigmaYesterday {
+  ticker: string;
+  name: string | null;
+  sigma: number;
+  weight: number;
+  marketCap: number;
+  stdLogReturn: number;
+  meanLogReturn: number;
+  sampleSize: number;
+  last: tendenzApiSigmaYesterdayDay;
+  secondLast: tendenzApiSigmaYesterdayDay;
+}
